Move TestFile to customer_test.dart
